Extract $consecutiveSiblingGroups into lexical_helper
It is a pure node-grouping utility with no dependency on Contents state, so it
belongs with the other Lexical node helpers.
